refactor: move Date extensions to TaskCell group path, add custom date formatter creating method
